requirements.txt
streamlit>=1.28.0
supabase>=1.0.0
python-dotenv==1.0.0
pandas>=2.0.0
audio-recorder-streamlit==0.0.9
SpeechRecognition==3.10.0
pydub==0.25.1
requests>=2.31.0
watchdog==3.0.0
deep-translator==1.11.4
pyaudioop
py-aifc